+++ gtest_parallel.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python2
 # Copyright 2013 Google Inc. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
